Move IoU loss to common
