+++ src/otx/algo/detection/losses/__init__.py
@@ -4,7 +4,6 @@
 """Custom OTX Losses for Object Detection."""
 
 from .atss_loss import ATSSCriterion
-from .iou_loss import IoULoss
 from .rtdetr_loss import DetrCriterion
 from .rtmdet_loss import RTMDetCriterion
 from .ssd_loss import SSDCriterion
@@ -13,7 +12,6 @@ from .yolox_loss import YOLOXCriterion
 
 __all__ = [
     "ATSSCriterion",
-    "IoULoss",
     "DetrCriterion",
     "RTMDetCriterion",
     "SSDCriterion",
